+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Add binder for nullable boolean options. (Shell)
 
 ### [0.1.0] - 2022-05-04